requirements.txt
opencv-python
pyautogui
tensorflow
numpy
Pillow